+++ lib/niconico.rb
@@ -114,5 +114,6 @@ require 'niconico/mylist'
 require 'niconico/ranking'
 require 'niconico/channel'
 require 'niconico/live'
+require 'niconico/live/client'
 require 'niconico/live/mypage'
 require 'niconico/nico_api'
