+++ lib/niconico/live/api.rb
@@ -1,12 +1,9 @@
 # coding: utf-8
 require 'time'
 require 'openssl'
+require 'niconico/live/util'
 
 class Niconico
-  def live_api
-    Live::API.new(self.agent)
-  end
-
   class Live
     class API
       class NoPublicKeyProvided < Exception; end
@@ -56,7 +53,7 @@ class Niconico
       end
 
       def get(id)
-        id = normalize_id(id)
+        id = Util::normalize_id(id)
 
         page = agent.get("http://live.nicovideo.jp/gate/#{id}")
 
@@ -102,7 +99,7 @@ class Niconico
       end
 
       def get_player_status(id, public_key = nil)
-        id = normalize_id(id)
+        id = Util::normalize_id(id)
         page = agent.get("http://ow.live.nicovideo.jp/api/getplayerstatus?locale=GLOBAL&lang=ja%2Djp&v=#{id}&seat%5Flocale=JP")
         if page.body[0] == 'c' # encrypted
           page = Nokogiri::XML(decrypt_encrypted_player_status(page.body, public_key))
@@ -185,11 +182,11 @@ class Niconico
 
       def watching_reservations
         page = agent.get(URL_WATCHINGRESERVATION_LIST)
-        page.search('vid').map(&:inner_text).map{ |_| normalize_id(_) }
+        page.search('vid').map(&:inner_text).map{ |_| Util::normalize_id(_) }
       end
 
       def accept_watching_reservation(id_)
-        id = normalize_id(id_, with_lv: false)
+        id = Util::normalize_id(id_, with_lv: false)
         page = agent.get("http://live.nicovideo.jp/api/watchingreservation?mode=confirm_watch_my&vid=#{id}&next_url&analytic")
         token = page.at('#reserve button')['onclick'].scan(/'(.+?)'/)[1][0]
 
@@ -222,40 +219,6 @@ class Niconico
 
         body = cipher.update(encrypted_body) + cipher.final
         body.force_encoding('utf-8')
-      end
-
-      def remove_timeshifts(ids)
-        post_body = "delete=timeshift&confirm=#{fetch_token}"
-        if ids.size == 0
-          return
-        end
-        ids.each do |id|
-          id = normalize_id(id, with_lv: false)
-          # mechanize doesn't support multiple values for the same key in query.
-          post_body += "&vid%5B%5D=#{id}"
-        end
-        agent.post(
-          'http://live.nicovideo.jp/my.php',
-          post_body,
-          'Content-Type' => 'application/x-www-form-urlencoded'
-        )
-      end
-
-      private
-
-      def fetch_token
-        page = agent.get('http://live.nicovideo.jp/my')
-        page.at('#confirm').attr('value')
-      end
-
-      def normalize_id(id, with_lv: true)
-        id = id.to_s
-
-        if with_lv
-          id.start_with?('lv') ? id : "lv#{id}"
-        else
-          id.start_with?('lv') ? id[2..-1] : id
-        end
       end
     end
   end
